build: python:3.11-slim image, buildx layer caching, drop tox
Self-contained multi-stage build (conda env, cbtemulator, slim prod with
nginx/supervisor/uwsgi) replaces the prebuilt tiangolo base; adds
override/nginx.conf, which the slim base doesn't ship.
Drop tox: it pulled cachetools/pluggy versions that conda-pack mixed with the
pins and broke the app import; retire the tox-only tox.ini and superseded
.travis.yml.
CI and the Docker Hub trigger build via Buildx with mode=max cache (gha and
registry) so the conda and go stages are reused across builds.
Fold the ingest-setup idempotency test into the module that already registers
the session emulator fixture; a second importer started a second emulator on
the same port and errored the suite.
