+++ cloudbuild.yaml
@@ -5,19 +5,24 @@ steps:
     args: ["-c", "docker login --username=$$USERNAME --password=$$PASSWORD"]
     secretEnv: ["USERNAME", "PASSWORD"]
 
+  # Create a buildx builder; the docker-container driver is required to export a
+  # registry-backed layer cache (the default docker driver cannot).
+  - name: "gcr.io/cloud-builders/docker"
+    entrypoint: "bash"
+    args: ["-c", "docker buildx create --name pcgbuilder --driver docker-container --use"]
+
+  # Build with a registry layer cache and push. mode=max caches intermediate
+  # stages (conda env, cbtemulator) so tagged rebuilds reuse them.
   - name: "gcr.io/cloud-builders/docker"
     entrypoint: "bash"
     args:
       - "-c"
       - |
-        docker build -t $$USERNAME/pychunkedgraph:$TAG_NAME .
-    timeout: 600s
-    secretEnv: ["USERNAME"]
-
-  # Push the final image to Dockerhub
-  - name: "gcr.io/cloud-builders/docker"
-    entrypoint: "bash"
-    args: ["-c", "docker push $$USERNAME/pychunkedgraph:$TAG_NAME"]
+        docker buildx build \
+          --cache-from type=registry,ref=$$USERNAME/pychunkedgraph:buildcache \
+          --cache-to type=registry,ref=$$USERNAME/pychunkedgraph:buildcache,mode=max \
+          --tag $$USERNAME/pychunkedgraph:$TAG_NAME \
+          --push .
     secretEnv: ["USERNAME"]
 
 availableSecrets:
@@ -26,3 +31,7 @@ availableSecrets:
       env: "PASSWORD"
     - versionName: projects/$PROJECT_ID/secrets/docker-username/versions/1
       env: "USERNAME"
+
+timeout: 3600s
+options:
+  machineType: "E2_HIGHCPU_8"
