+++ override/nginx.conf
@@ -1,0 +1,43 @@
+user  nginx;
+worker_processes  1;
+
+error_log  /var/log/nginx/error.log warn;
+pid        /var/run/nginx.pid;
+
+daemon off;
+
+events {
+    worker_connections  1024;
+}
+
+http {
+    include       /etc/nginx/mime.types;
+    default_type  application/octet-stream;
+
+    log_format  main  '$remote_addr - $remote_user [$time_local] "$request" '
+                      '$status $body_bytes_sent "$http_referer" '
+                      '"$http_user_agent" "$http_x_forwarded_for"';
+
+    map $http_user_agent $loggable {
+        default 1;
+        ~*kube-probe 0;
+    }
+
+    access_log  /var/log/nginx/access.log  main  if=$loggable;
+
+    sendfile        on;
+
+    keepalive_timeout  65;
+
+    client_max_body_size 0;
+
+    include /etc/nginx/conf.d/*.conf;
+
+    server {
+        listen 80;
+        location / {
+            include uwsgi_params;
+            uwsgi_pass unix:///tmp/uwsgi.sock;
+        }
+    }
+}
