+++ requirements.yml
@@ -4,7 +4,6 @@ channels:
 dependencies:
   - python==3.11.4
   - pip
-  - tox
   - uwsgi==2.0.21
   - graph-tool-base==2.58
   - zstandard==0.19.0  # ugly hack to force PyPi install 0.21.0
