--- tox.ini
@@ -1,13 +1,0 @@
-[tox]
-envlist = py311
-requires = tox-conda
-
-[testenv]
-setenv = HOME = {env:HOME}
-
-deps = pytest
-       pytest-cov
-       pytest-mock
-       pytest-timeout
-conda_env = requirements.yml
-commands = python -m pytest -v {posargs} ./pychunkedgraph/tests/
